move resolveInfo
