+++ libs/gi/page-team/src/CharacterDisplay/Tabs/TabOptimize/Components/OptimizationTargetEditorList.tsx
@@ -4,9 +4,8 @@ import {
 } from '@genshin-optimizer/common/ui'
 import { objPathValue } from '@genshin-optimizer/common/util'
 import type { StatFilterSetting, StatFilters } from '@genshin-optimizer/gi/db'
-import { DataContext } from '@genshin-optimizer/gi/ui'
+import { DataContext, resolveInfo } from '@genshin-optimizer/gi/ui'
 import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { resolveInfo } from '@genshin-optimizer/gi/uidata'
 import {
   CheckBox,
   CheckBoxOutlineBlank,
