+++ libs/gi/page-team/src/CharacterDisplay/Tabs/TabOptimize/Components/OptimizationTargetSelector.tsx
@@ -2,9 +2,12 @@ import { useBoolState } from '@genshin-optimizer/common/react-util'
 import { ImgIcon, SqBadge } from '@genshin-optimizer/common/ui'
 import { objPathValue } from '@genshin-optimizer/common/util'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
-import { DataContext, getDisplayHeader } from '@genshin-optimizer/gi/ui'
+import {
+  DataContext,
+  getDisplayHeader,
+  resolveInfo,
+} from '@genshin-optimizer/gi/ui'
 import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { resolveInfo } from '@genshin-optimizer/gi/uidata'
 import { Box, Button, Divider, Stack } from '@mui/material'
 import { useCallback, useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
