+++ libs/gi/page-team/src/CharacterDisplay/Tabs/TabOptimize/index.tsx
@@ -57,14 +57,12 @@ import {
   OptimizationTargetContext,
   ReactionToggle,
   getTeamData,
+  resolveInfo,
   useGlobalError,
   useTeamData,
 } from '@genshin-optimizer/gi/ui'
-import {
-  resolveInfo,
-  uiDataForTeam,
-  type UIData,
-} from '@genshin-optimizer/gi/uidata'
+import type { UIData } from '@genshin-optimizer/gi/uidata'
+import { uiDataForTeam } from '@genshin-optimizer/gi/uidata'
 import type { NumNode } from '@genshin-optimizer/gi/wr'
 import { mergeData, optimize } from '@genshin-optimizer/gi/wr'
 import {
