+++ libs/gi/ui/src/components/character/editor/OptimizationTargetDisplay.tsx
@@ -2,7 +2,6 @@ import { CardThemed, ImgIcon, SqBadge } from '@genshin-optimizer/common/ui'
 import { objPathValue } from '@genshin-optimizer/common/util'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
 import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { resolveInfo } from '@genshin-optimizer/gi/uidata'
 import TrackChangesIcon from '@mui/icons-material/TrackChanges'
 import {
   Box,
@@ -15,7 +14,7 @@ import {
 import { useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { DataContext } from '../../../context'
-import { getDisplayHeader } from '../../../util'
+import { getDisplayHeader, resolveInfo } from '../../../util'
 export function OptimizationTargetDisplay({
   optimizationTarget,
   showEmptyTargets = false,
