+++ libs/gi/ui/src/components/weapon/WeaponCardPico.tsx
@@ -5,12 +5,12 @@ import { useWeapon } from '@genshin-optimizer/gi/db-ui'
 import { getWeaponSheet } from '@genshin-optimizer/gi/sheets'
 import { getWeaponStat, weaponHasRefinement } from '@genshin-optimizer/gi/stats'
 import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { computeUIData, resolveInfo } from '@genshin-optimizer/gi/uidata'
+import { computeUIData } from '@genshin-optimizer/gi/uidata'
 import { getLevelString } from '@genshin-optimizer/gi/util'
 import { dataObjForWeapon, uiInput as input } from '@genshin-optimizer/gi/wr'
 import { Box, Typography } from '@mui/material'
 import { useMemo } from 'react'
-import { getCalcDisplay } from '../../util'
+import { getCalcDisplay, resolveInfo } from '../../util'
 import { WeaponNameTooltip } from './WeaponNameTooltip'
 
 export function WeaponCardPico({ weaponId }: { weaponId: string }) {
