+++ libs/gi/ui/src/util/index.ts
@@ -3,4 +3,5 @@ export * from './getCalcDisplay'
 export * from './getDisplayUtil'
 export * from './iconAsset'
 export * from './isDev'
+export * from './resolveInfo'
 export * from './weaponSortUtil'
