--- libs/gi/uidata/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@nx/react/babel",
-      {
-        "runtime": "automatic",
-        "useBuiltIns": "usage",
-        "importSource": "@emotion/react"
-      }
-    ]
-  ],
-  "plugins": ["@emotion/babel-plugin"]
-}
