+++ libs/gi/uidata/README.md
@@ -1,7 +1,3 @@
 # gi-uidata
 
 This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test gi-uidata` to execute the unit tests via [Jest](https://jestjs.io).
